# docs/api/api.yaml
openapi: 3.0.0
info:
  title: Wutsi Site API
  description: |
    This API manages the different sites manages by the Wutsi Platform.
    The API offer the following functionalities:
    - Registering new sites
    - Updating site informations
    - Managing the configuration of each sites
    This API give to the Platform the capability of multi-tenancy.
  version: 1.0.0
servers:
  - url: 'https://wutsi-site-test.herokuapp.com'
    description: Sandbox
  - url: 'https://wutsi-site-prod.herokuapp.com'
    description: Production
paths:
  '/v1/sites/{id}':
    get:
      operationId: get
      summary: Return the information about a given site
      parameters:
        - in: path
          name: id
          description: ID of the site
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSiteResponse'
      security:
        - api_key:
            - site-read
      tags:
        - Site
    post:
      operationId: update
      summary: Update a Site
      description: |
        Update a site information
      parameters:
        - in: path
          name: id
          description: ID of the site
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSiteResponse'
      security:
        - api_key:
            - site-manage
      tags:
        - Site
  '/v1/sites/{id}/attributes/{urn}':
    post:
      operationId: setAttribute
      summary: Update a site attribute
      description: |
        Update a site attribute.
        After the attribute is updated, the event `urn:event:wutsi:site:updated` will be fired.
      parameters:
        - in: path
          name: id
          description: ID of the site
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: urn
          description: URN of the attribute
          required: true
          schema:
            type: string
      requestBody:
        description: request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAttributeRequest'
      responses:
        '200':
          description: Success
      security:
        - api_key:
            - site-manage
      tags:
        - Site
  /v1/sites:
    get:
      operationId: search
      summary: Search sites
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchSiteResponse'
      security:
        - api_key:
            - site-read
      tags:
        - Site
    post:
      operationId: create
      summary: Create a new Site
      description: |
        Create a new site.
        After the site is created, the event `urn:event:wutsi:site:created` will be fired.
      requestBody:
        description: Site request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSiteResponse'
      security:
        - api_key:
            - site-manage
      tags:
        - Site
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    SearchSiteResponse:
      type: object
      properties:
        sites:
          type: array
          items:
            $ref: '#/components/schemas/SiteSummary'
    SiteSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: ZeSite
        displayName:
          type: string
          example: ZeSite
    GetSiteResponse:
      type: object
      properties:
        site:
          type: object
          $ref: '#/components/schemas/Site'
    Site:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: ZeSite
        domainName:
          type: string
          example: zesite.com
        displayName:
          type: string
          example: ZeSite
        websiteUrl:
          type: string
          format: url
          example: 'https://www.zesite.com'
        language:
          type: string
          format: url
          example: fr
          description: Default language code of the site. The code is in the ISO0639-1 format
        currency:
          type: string
          format: url
          example: XFA
          description: Default currency code of the site. The code is in the ISO-4217 format
        internationalCurrency:
          type: string
          format: url
          example: EUR
          description: Currency to use for international customers. The code is in the ISO-4217 format
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
    Attribute:
      type: object
      description: |
        Attribute of a site.
        Attributes are used to store the site configuration
      properties:
        urn:
          type: string
          nullable: false
          description: |
            URN of the attribute.
            The format should be urn:attribute:wutsi:<domain>:<name> where:
            - <domain> is the application domain
            - <name> is the name of the attribute
          example: 'urn:attribute:wutsi:test:foo'
        value:
          type: string
          description: Value of the attribute
          nullable: false
          example: bar
    UpdateSiteRequest:
      type: object
      properties:
        name:
          type: string
          example: ZeSite
        domainName:
          type: string
          example: zesite.com
        displayName:
          type: string
          example: ZeSite
        language:
          type: string
          format: url
          example: fr
        currency:
          type: string
          format: url
          example: XFA
        internationalCurrency:
          type: string
          format: url
          example: EUR
      required:
        - name
        - domainName
        - language
        - currency
        - internationalCurrency
    UpdateSiteResponse:
      type: object
      properties:
        siteId:
          type: integer
          format: int64
    SetAttributeRequest:
      type: object
      properties:
        value:
          type: string
          nullable: true
    CreateSiteRequest:
      type: object
      properties:
        name:
          type: string
          example: ZeSite
        domainName:
          type: string
          example: zesite.com
        displayName:
          type: string
          example: ZeSite
        language:
          type: string
          format: url
          example: fr
        currency:
          type: string
          format: url
          example: XFA
        internationalCurrency:
          type: string
          format: url
          example: EUR
      required:
        - name
        - domainName
        - language
        - currency
        - internationalCurrency
    CreateSiteResponse:
      type: object
      properties:
        siteId:
          type: integer
          format: int64